backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
opencv-python==4.10.0.84
openai==1.57.2
requests==2.32.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
numpy>=1.24.0
Pillow>=10.0.0
soundfile==0.12.1
librosa==0.10.2
python-multipart==0.0.12
Werkzeug==3.0.6
scikit-learn==1.5.2
urllib3>=1.26.0
certifi>=2023.7.22

# Optional ML dependencies (comment out if having SSL issues)
sentence-transformers==3.3.1
transformers==4.46.2
torch>=2.6.0
